+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 ## Changelog
 
+## [2.3.1](https://github.com/logdna/logdna-browser/compare/v2.3.0...v2.3.1) (2024-04-24)
+
+
+### Bug Fixes
+
+* **logger**: Updates for stacktrace handling [6a14611](https://github.com/logdna/logdna-browser/commit/6a14611403c3e63c7dd2db4c15ef045154238196) - Terry Moore
+
+# [2.3.0](https://github.com/logdna/logdna-browser/compare/v2.2.0...v2.3.0) (2024-04-08)
+
+
+### Features
+
+* **logger**: Updates for better stacktrace handling [d7ba2b7](https://github.com/logdna/logdna-browser/commit/d7ba2b75cc522efe829f96c1fcc5d7ae0e476447) - Terry Moore
+
+# [2.2.0](https://github.com/logdna/logdna-browser/compare/v2.1.0...v2.2.0) (2024-03-27)
+
+
+### Features
+
+* **logger**: Additional internal logger options [f79fd80](https://github.com/logdna/logdna-browser/commit/f79fd804f8a2b32a5e85252e62af9da5050aa6da) - Terry Moore
+
 # [2.1.0](https://github.com/logdna/logdna-browser/compare/v2.0.11...v2.1.0) (2022-11-28)